Rename panels.rs to parameter_panel.rs to match its content.
