+++ crates/nodebox-gui/src/lib.rs
@@ -33,7 +33,7 @@ mod network_view;
 mod node_library;
 mod node_selection_dialog;
 mod pan_zoom;
-mod panels;
+mod parameter_panel;
 pub mod render_worker;
 pub mod state;
 mod theme;
